Add a tool to visulize the generated tfrecored from preprocessing pipeline
